onnx_runtime_server/experiment2/client/pyproject.toml
[tool.poetry]
name = "client"
version = "0.1.0"
description = ""
authors = ["YuyaK-0913 <kuniyu.0913@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
googleapis-common-protos = "^1.65.0"
protobuf = "==3.20.*"
requests = "^2.32.3"
numpy = "^1.21"
grpcio-tools = "^1.43"
grpcio = "^1.66.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
